Refactoring routingURls
