+++ _api/webhooks/neynar_mention.ts
@@ -3,12 +3,12 @@ import type { VercelRequest, VercelResponse } from '@vercel/node';
 import { getGiveBotProfileId } from '../../api-lib/colinks/helperAccounts.ts';
 import { IS_LOCAL_ENV } from '../../api-lib/config';
 import { fetchViewerInfo } from '../../api-lib/frames/give/fetchViewerInfo.tsx';
+import { FRAME_ROUTER_URL_BASE } from '../../api-lib/frames/routingUrls.ts';
 import { insertInteractionEvents } from '../../api-lib/gql/mutations';
 import { errorResponse } from '../../api-lib/HttpError';
 import { publishCast } from '../../api-lib/neynar';
 import { findOrCreateProfileByFid } from '../../api-lib/neynar/findOrCreateProfileByFid.ts';
 import { isValidSignature } from '../../api-lib/neynarSignature';
-import { webAppURL } from '../../src/config/webAppURL.ts';
 import {
   checkPointsAndCreateGive,
   fetchPoints,
@@ -191,7 +191,6 @@ const parseSkill = (text: string) => {
   return skillMatch ? skillMatch[1] : undefined;
 };
 
-const FRAME_ROUTER_URL_BASE = `${webAppURL('colinks')}/api/frames/router`;
 export const getFrameUrl = (frameId: string, resourceId?: number) => {
   let url = `${FRAME_ROUTER_URL_BASE}/meta/${frameId}`;
